requirements.txt
fastapi
uvicorn
openai
tweepy
python-dotenv
textblob
